Fix compile warnings.

+++ src/conio/cterm.h
@@ -90,7 +90,7 @@ extern "C" {
 extern struct cterminal cterm;
 
 void cterm_init(int height, int width, int xpos, int ypos, int backlines, unsigned char *scrollback);
-char *cterm_write(unsigned char *buf, int buflen, char *retbuf, int retsize, int *speed);
+char *cterm_write(unsigned char *buf, int buflen, char *retbuf, size_t retsize, int *speed);
 int cterm_openlog(char *logfile, int logtype);
 void cterm_closelog(void);
 void cterm_end(void);
